+++ packages/core/src/constructs.ts
@@ -3575,6 +3575,9 @@ export function merge(
       state: DocState<Record<string | symbol, unknown>>,
       _defaultValue?,
     ) {
+      let brief: Message | undefined;
+      let description: Message | undefined;
+      let footer: Message | undefined;
       const fragments = parsers.flatMap((p, i) => {
         let parserState: DocState<unknown>;
 
@@ -3608,7 +3611,11 @@ export function merge(
         // due to the way merge() handles disparate parser types.
         // The runtime logic ensures we are passing the correct state slice or unavailable.
         // deno-lint-ignore no-explicit-any
-        return p.getDocFragments(parserState as any, undefined).fragments;
+        const docFragments = p.getDocFragments(parserState as any, undefined);
+        brief ??= docFragments.brief;
+        description ??= docFragments.description;
+        footer ??= docFragments.footer;
+        return docFragments.fragments;
       });
       const entries: DocEntry[] = fragments.filter((f) => f.type === "entry");
       const sections: DocSection[] = [];
@@ -3626,6 +3633,9 @@ export function merge(
         const labeledSection: DocSection = { title: label, entries };
         sections.push(labeledSection);
         return {
+          brief,
+          description,
+          footer,
           fragments: sections.map<DocFragment>((s) => ({
             ...s,
             type: "section",
@@ -3634,6 +3644,9 @@ export function merge(
       }
 
       return {
+        brief,
+        description,
+        footer,
         fragments: [
           ...sections.map<DocFragment>((s) => ({ ...s, type: "section" })),
           { type: "section", entries },